requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
stripe~=7.6.0
python-dotenv~=1.0.0
django-filter~=23.4